requirements.txt
av==10.0.0
numpy==1.23.2
opencv_python_headless==4.6.0.66
Pillow==9.3.0
streamlit==1.12.0
streamlit_webrtc==0.44.0
